+++ test/media_player/test_async_setup_entry.py
@@ -17,7 +17,6 @@ TEST_MODULE = "custom_components.spotcast.media_player"
 
 class TestMediaPlayerSetup(IsolatedAsyncioTestCase):
 
-    @patch(f"{TEST_MODULE}.async_track_time_interval")
     @patch.object(DeviceManager, "async_update")
     @patch.object(
         SpotifyAccount,
@@ -28,18 +27,15 @@ class TestMediaPlayerSetup(IsolatedAsyncioTestCase):
         self,
         mock_account: AsyncMock,
         mock_update: AsyncMock,
-        mock_track_time: AsyncMock,
     ):
 
         self.mock_update = mock_update
-        self.mock_track_time = mock_track_time
 
         self.mock_hass = MagicMock(spec=HomeAssistant)
         self.mock_entry = MagicMock(spec=ConfigEntry)
         self.mock_callback = MagicMock(spec=AddEntitiesCallback)
         self.mock_entry.entry_id = "12345"
         self.mock_hass.data = {}
-        self.mock_track_time.return_value = "foo"
 
         await async_setup_entry(
             self.mock_hass,
@@ -53,14 +49,6 @@ class TestMediaPlayerSetup(IsolatedAsyncioTestCase):
         except AssertionError:
             self.fail()
 
-    async def test_async_track_time_called_properly(self):
-        try:
-            self.mock_track_time.assert_called()
-        except AssertionError:
-            self.fail()
-
-    async def test_listener_registered_for_unload(self):
-        try:
-            self.mock_entry.async_on_unload.assert_called_with("foo")
-        except AssertionError:
-            self.fail()
+    async def test_device_manager_stored_in_hass_data(self):
+        device_manager = self.mock_hass            .data["spotcast"]["12345"]["device_manager"]
+        self.assertIsInstance(device_manager, DeviceManager)