+++ test/sensor/abstract_sensor/test_state.py
@@ -4,21 +4,23 @@ from unittest import TestCase
 from unittest.mock import MagicMock
 
 from custom_components.spotcast.sensor.abstract_sensor import SpotcastSensor
+from custom_components.spotcast.coordinator import SpotcastCoordinator
 from custom_components.spotcast.spotify import SpotifyAccount
 
 
 class DummySensor(SpotcastSensor):
     UNITS_OF_MEASURE = "foos"
 
-    async def _async_update_process(self):
+    def _update_from_coordinator(self):
         ...
 
 
 class TestSensorState(TestCase):
 
     def setUp(self):
-        self.account = MagicMock(spec=SpotifyAccount)
-        self.sensor = DummySensor(self.account)
+        self.coordinator = MagicMock(spec=SpotcastCoordinator)
+        self.coordinator.account = MagicMock(spec=SpotifyAccount)
+        self.sensor = DummySensor(self.coordinator)
         self.sensor._attr_state = 10
 
     def test_sensor_state_returned(self):