+++ test/sensor/abstract_sensor/test_units_of_measurement.py
@@ -4,21 +4,23 @@ from unittest import TestCase
 from unittest.mock import MagicMock
 
 from custom_components.spotcast.sensor.abstract_sensor import SpotcastSensor
+from custom_components.spotcast.coordinator import SpotcastCoordinator
 from custom_components.spotcast.spotify import SpotifyAccount
 
 
 class DummySensor(SpotcastSensor):
     UNITS_OF_MEASURE = "foos"
 
-    async def _async_update_process(self):
+    def _update_from_coordinator(self):
         ...
 
 
 class TestUnitsMeasurement(TestCase):
 
     def setUp(self):
-        self.account = MagicMock(spec=SpotifyAccount)
-        self.sensor = DummySensor(self.account)
+        self.coordinator = MagicMock(spec=SpotcastCoordinator)
+        self.coordinator.account = MagicMock(spec=SpotifyAccount)
+        self.sensor = DummySensor(self.coordinator)
 
     def test_units_of_measurements_from_contstants(self):
         self.assertEqual(self.sensor.unit_of_measurement, "foos")